+++ iti-react/Hooks/index.ts
@@ -1,3 +1,2 @@
-﻿export * from './ChildValidChange'
-export * from './UseReadiness'
+﻿export * from './UseReadiness'
 export * from './UsePaginationHelpers'