+++ iti-react/__tests__/ValidatedSelect.test.ts
@@ -1,0 +1,3 @@
+it('does not unnecessarily rerender', () => {
+    fail()
+})