+++ test-website/src/_Redux/Auth/AuthSelectors.ts
@@ -1,14 +1,12 @@
 ﻿import { UserDto } from 'Models'
 import { AppState } from '../AppState'
 import { nullToUndefined } from '@interface-technologies/iti-react'
+import { RequestStatus } from '_Redux/Common/RequestStatus'
 
 export function userSelector(state: AppState): UserDto | undefined {
     return nullToUndefined(state.auth.user)
 }
 
-export function requiredUserSelector(state: AppState): UserDto {
-    const user = userSelector(state)
-    if (!user) throw new Error('User is null or undefined.')
-
-    return user
+export const authSelectors = {
+    logInRequestStatus: (state: AppState): RequestStatus => state.auth.logInRequestStatus
 }